Move internal features out of app module
